Remove DOM dependencies from background.js
Chrome's manifest v3 does not support DOM API in the background / for
service workers. This commit separates "foreground" functions in utils/location to utils/currentLocation. Said functions sometimes requires access to the DOM to function properly.

Additionaly, this commit tweaks a few module imports to avoid having the entirity of environment/foreground inserted into background.js.
