+++ lib/core/migrate/migrate.js
@@ -1,7 +1,7 @@
 /* @flow */
 
 import { takeRightWhile, isObject, mapKeys, uniqBy } from 'lodash-es';
-import { Storage } from '../../environment';
+import * as Storage from '../../environment/foreground/storage';
 import { forEachSeq } from '../../utils/array';
 import { execRegexes } from '../../utils/location';
 import * as Modules from '../modules/storage';
