+++ lib/core/migrate/migrators.js
@@ -1,7 +1,7 @@
 /* @flow */
 
 import { isEqual } from 'lodash-es';
-import { Storage } from '../../environment';
+import * as Storage from '../../environment/foreground/storage';
 import * as Options from '../options/storage';
 
 export async function updateOption(moduleID: string, optionName: string, formerDefaultValue: mixed, valueOrFunction: mixed) {
