+++ lib/core/modules/storage.js
@@ -1,6 +1,6 @@
 /* @flow */
 
-import { Storage } from '../../environment';
+import * as Storage from '../../environment/foreground/storage';
 
 export const storage = Storage.wrapBlob('RES.modulePrefs',
 	(): boolean => { throw new Error('Default module enabled state should never be accessed'); });
