+++ lib/core/options/storage.js
@@ -1,6 +1,6 @@
 /* @flow */
 
-import { Storage } from '../../environment';
+import * as Storage from '../../environment/foreground/storage';
 import { getModuleId } from '../module';
 import type { OpaqueModuleId } from '../module';
