+++ lib/modules/quarantineHide.js
@@ -2,8 +2,7 @@
 
 import $ from 'jquery';
 import { Module } from '../core/module';
-import { BodyClasses, frameThrottle, watchForThings } from '../utils';
-import { inQuarantinedSubreddit } from '../utils/location';
+import { BodyClasses, frameThrottle, inQuarantinedSubreddit, watchForThings } from '../utils';
 
 export const module: Module<*> = new Module('quarantineHide');
