+++ lib/utils/Thing.js
@@ -5,7 +5,8 @@ import { filterMap } from './array';
 import { throttleQueuePositionReset } from './async';
 import { click, getPercentageVisibleYAxis, scrollToElement } from './dom';
 import { downcast } from './flow';
-import { currentSubreddit, regexes } from './location';
+import { currentSubreddit } from './currentLocation';
+import { regexes } from './location';
 
 const elementMap = new WeakMap();
 const things = new Set();
