+++ lib/utils/createElement.js
@@ -6,7 +6,7 @@ import type { Iteratee } from './array';
 import { waitForEvent } from './dom';
 import { downcast } from './flow';
 import { addFloater } from './floater';
-import { isCurrentSubreddit, currentSubreddit } from './location';
+import { isCurrentSubreddit, currentSubreddit } from './currentLocation';
 import * as string from './string';
 
 export function toggleButton(
