+++ lib/utils/floater.js
@@ -4,7 +4,7 @@ import { memoize } from 'lodash-es';
 import { getD2xBodyOffset, getHeaderOffset } from './dom';
 import * as PagePhases from './pagePhases';
 import { frameThrottle } from './async';
-import { isAppType } from './location';
+import { isAppType } from './currentLocation';
 import { html } from './string';
 
 class Container {
