+++ lib/utils/index.js
@@ -114,24 +114,27 @@ export {
 } from './localization';
 
 export {
+	execRegexes,
+	isEmptyLink,
+	fullLocation,
+	isCommentCode,
+	regexes,
+} from './location';
+
+export {
 	appType,
 	currentDomain,
 	currentMultireddit,
 	currentSubreddit,
 	currentUserProfile,
-	execRegexes,
-	isAppType,
-	isCommentCode,
+	inQuarantinedSubreddit,
 	isCurrentMultireddit,
 	isCurrentSubreddit,
-	isEmptyLink,
-	isPageType,
 	matchesPageLocation,
-	matchesPageRegex,
-	fullLocation,
+	isAppType,
+	isPageType,
 	pageType,
-	regexes,
-} from './location';
+} from './currentLocation';
 
 export {
 	numericalCompare,
