+++ lib/utils/location.js
@@ -1,21 +1,5 @@
 /* @flow */
 
-import { once, partition, memoize } from 'lodash-es';
-import { waitForEvent } from './dom';
-
-export function matchesPageLocation(includes: Array<PageType | AppType | RegExp>, excludes: Array<PageType | AppType | RegExp> = []): boolean {
-	const [includeStrings, includeRegExps]: any = partition(includes, (x: string | RegExp): boolean => typeof x === 'string');
-	const [excludeStrings, excludeRegExps]: any = partition(excludes, (x: string | RegExp): boolean => typeof x === 'string');
-
-	return (
-		!excludes.length ||
-		!(isPageType(...excludeStrings) || isAppType(...excludeStrings) || matchesPageRegex(...excludeRegExps))
-	) && (
-		!includes.length ||
-		(isPageType(...includeStrings) || isAppType(...includeStrings) || matchesPageRegex(...includeRegExps))
-	);
-}
-
 /* eslint-disable key-spacing */
 export const regexes: { [string]: RegExp } = {
 	frontpage:        /^\/(?:hot|new|rising|controversial|top)?(?:\/|$)/i,
@@ -55,22 +39,7 @@ export const execRegexes: { [string]: (path: string) => ?string[] } = {
 export type PageType = 'wiki' | 'search' | 'profile' | 'profile2x' | 'profileCommentsPage' | 'stylesheet' | 'modqueue' | 'subredditAbout' | 'comments' | 'commentsLinklist' | 'liveThread' | 'inbox' | 'submit' | 'account' | 'prefs' | 'linklist';
 export type AppType = 'r2' | 'd2x' | 'options';
 
-export const appType = once((): AppType => {
-	if (document.documentElement.hasAttribute('res-options')) {
-		return 'options';
-	}
-	if (document.documentElement.getAttribute('xmlns')) {
-		return 'r2';
-	}
-	return 'd2x';
-});
-
-export function isAppType(...types: AppType[]): boolean {
-	const thisApp = appType();
-	return types.some(type => type === thisApp);
-}
-
-const appPageTypes: { [AppType]: {| default?: PageType, pageTypes: PageType[] |} } = {
+export const appPageTypes: { [AppType]: {| default?: PageType, pageTypes: PageType[] |} } = {
 	r2: {
 		default: 'linklist',
 		pageTypes: ['wiki', 'search', 'stylesheet', 'modqueue', 'subredditAbout', 'comments', 'commentsLinklist', 'profile', 'liveThread', 'inbox', 'submit', 'account', 'prefs'],
@@ -82,55 +51,6 @@ const appPageTypes: { [AppType]: {| default?: PageType, pageTypes: PageType[] |}
 		pageTypes: [],
 	},
 };
-
-export const pageType = memoize((): ?PageType => {
-	waitForEvent(document, 'reddit.urlChanged').then(() => {
-		pageType.cache.clear();
-	});
-	const spec = appPageTypes[appType()];
-	return spec.pageTypes.find(pageType => regexes[pageType].test(location.pathname)) || spec.default;
-});
-
-export function isPageType(...types: PageType[]): boolean {
-	const thisPage = pageType();
-	return types.some(type => type === thisPage);
-}
-
-export function matchesPageRegex(...regexps: RegExp[]): boolean {
-	return regexps.some(regex => regex.test(location.pathname));
-}
-
-export const currentSubreddit = once((): string | void => {
-	const match = location.pathname.match(regexes.subreddit);
-	if (match) return match[1];
-});
-
-export function isCurrentSubreddit(...subreddits: string[]): boolean {
-	const sub = (currentSubreddit() || '').toLowerCase();
-	if (!sub) return false;
-	return subreddits.some(v => v.toLowerCase() === sub);
-}
-
-export const currentMultireddit = once((): string | void => {
-	const match = location.pathname.match(regexes.multireddit);
-	if (match) return match[1];
-});
-
-export function isCurrentMultireddit(...multireddits: string[]): boolean {
-	const multi = (currentMultireddit() || '').toLowerCase();
-	if (!multi) return false;
-	return multireddits.some(v => v.toLowerCase() === multi);
-}
-
-export const currentDomain = once((): string | void => {
-	const match = location.pathname.match(regexes.domain);
-	if (match) return match[1];
-});
-
-export const currentUserProfile = once((): string | void => {
-	const match = location.pathname.match(regexes.profile);
-	if (match) return match[1];
-});
 
 export const fullLocation = (pathname: string = location.pathname) => {
 	const regex = Object.keys(regexes).find(key => pathname.match(regexes[key]));
@@ -177,5 +97,3 @@ export function isEmptyLink(link: HTMLAnchorElement): boolean {
 	const href = link.getAttribute('href');
 	return typeof href !== 'string' || href.startsWith('javascript:') || href === '#'; // eslint-disable-line no-script-url
 }
-
-export const inQuarantinedSubreddit = once(() => document.body.classList.contains('quarantine'));
