+++ lib/utils/user.js
@@ -4,7 +4,8 @@ import { once } from 'lodash-es';
 import { ajax } from '../environment';
 import type { CurrentRedditUser, RedditAccount } from '../types/reddit';
 import { MINUTE } from './time';
-import { isAppType, regexes } from './location';
+import { regexes } from './location';
+import { isAppType } from './currentLocation';
 import * as string from './string';
 
 export const isLoggedIn = once(() => {
