+++ lib/utils/watchers.js
@@ -2,7 +2,7 @@
 
 import { memoize, once } from 'lodash-es';
 import { Thing } from './Thing';
-import { isPageType } from './location';
+import { isPageType } from './currentLocation';
 import { isLastNodeInDOM, watchForFutureChildren, watchForFutureDescendants, waitForAttach, waitForDescendantChange, waitForSelectorMatch, watchForChildren } from './dom';
 import { contentLoaded } from './pagePhases';
 import { MINUTE } from './time';
